feat: Monitor Peserta
WIP VIEW ON ADI

+++ app/Http/Controllers/AuthenticationController.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\AuthenticationRequest;
 use Carbon\Carbon;
 use DB;
-use Illuminate\Http\Request;
 use Str;
 
 class AuthenticationController extends Controller
@@ -21,19 +20,27 @@ class AuthenticationController extends Controller
         $email = $request->email;
         $password = $request->password;
 
-        if ($email === config('app.admin_email') && $password === config('app.admin_password')) {
+        if (
+            ($email === config('app.admin_email') && $password === config('app.admin_password')) ||
+            ($email === config('app.monitor_email') && $password === config('app.monitor_password'))
+        ) {
             $token = Str::random(60);
             $exp = Carbon::now()->addDay();
             if ($request->remember_me) {
                 $exp = Carbon::now()->addDays(7);
             }
 
-            session(['admin_logged_in' => true, 'admin_token' => $token]);
             DB::table('personal_access_tokens')->updateOrInsert(
                 ['token' => $token],
                 ['expiration' => $exp]
             );
 
+            if ($email === config('app.monitor_email')) {
+                session(['login_is' => 'monitor', 'login_uts_token' => $token]);
+                return to_route('monitor.index');
+            }
+
+            session(['login_is' => 'admin', 'login_uts_token' => $token]);
             return to_route('admin.dashboard');
         }
 
@@ -43,14 +50,15 @@ class AuthenticationController extends Controller
 
     public function logout()
     {
-        try{
-            $token = session('admin_token');
+        try {
+            $token = session('login_uts_token');
+            if(!$token) return to_route('kehadiran.index');
             DB::table('personal_access_tokens')->where(['token' => $token])->delete();
-            session()->forget('admin_token');
-            session()->forget('admin_logged_in');
+            session()->forget('login_uts_token');
+            session()->forget('login_is');
 
             return to_route('kehadiran.index');
-        }catch(\Exception $e){
+        } catch (\Exception $e) {
             flash()->error($e->getMessage());
             return back();
         }
